+++ CHANGELOG.md
@@ -12,6 +12,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * `include` directive
 
 
+### Fixed
+
+* Accept comment on `option` directive
+
+
 ## [1.0.0-alpha.3] - 2023-05-24
 
 ### Breaking changes